Add files via upload
Charming Thyme Trattoria is an extremely famous fine-dining restaurant in Kuala Lumpur, Malaysia. The restaurant always strives to provide the best fine-dining cuisine in the Kuala Lumpur area. Due to popular demand, guests are required to make a reservation to enjoy a meal at the restaurant. Hence, the goal of the project is to develop a restaurant management system for the restaurant using the Python 3 programming language. The focus of the management system is to process all the reservation data. However, it also
has the option to generate random dish recommendations if the guests are looking for an adventure and request to be surprised.